Fix invoke print-available-images and promote commands
Only 100 images are returned by ecr.list_images,
so it is possible to miss some of the recent images.

Now all the images are collected first, then 10 images of the sorted set of
image are returned.

The promote command has now an option to set the platform. Default is linux/amd64.